+++ backend/packages/integrations/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cat-factory/integrations
 
+## 0.153.12
+
+### Patch Changes
+
+- Updated dependencies [19baddf]
+  - @cat-factory/kernel@0.284.0
+
+## 0.153.11
+
+### Patch Changes
+
+- Updated dependencies [31f43c1]
+  - @cat-factory/contracts@0.290.0
+  - @cat-factory/kernel@0.283.0
+
 ## 0.153.10
 
 ### Patch Changes